requirements.txt
python-dotenv==1.0.1
pydantic==2.9.2
pandas==2.2.2
MetaTrader5==5.0.45
python-telegram-bot==21.4
numpy==1.26.4
backtesting==0.6.5
beautifulsoup4==4.12.3
streamlit==1.40.0